refactor: move tradeStateFromUrl to atom, to update it in a single place
