+++ apps/cowswap-frontend/src/modules/trade/hooks/useResetRecipient.ts
@@ -10,17 +10,17 @@ import { useDerivedTradeState } from './useDerivedTradeState'
 
 import { useIsAlternativeOrderModalVisible } from '../state/alternativeOrder'
 
-
 export function useResetRecipient(onChangeRecipient: (recipient: string | null) => void): null {
   const isAlternativeOrderModalVisible = useIsAlternativeOrderModalVisible()
   const tradeState = useDerivedTradeState()
   const tradeStateFromUrl = useTradeStateFromUrl()
   const postHooksRecipientOverride = usePostHooksRecipientOverride()
+  const hasTradeState = !!tradeStateFromUrl
   const { chainId } = useWalletInfo()
 
   const prevPostHooksRecipientOverride = usePrevious(postHooksRecipientOverride)
   const recipient = tradeState?.recipient
-  const hasRecipientInUrl = !!tradeStateFromUrl.recipient
+  const hasRecipientInUrl = !!tradeStateFromUrl?.recipient
 
   /**
    * Reset recipient value only once at App start if it's not set in URL
@@ -30,7 +30,7 @@ export function useResetRecipient(onChangeRecipient: (recipient: string | null) 
       onChangeRecipient(null)
     }
     // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [])
+  }, [hasTradeState])
 
   /**
    * Reset recipient whenever chainId changes
