+++ apps/cowswap-frontend/src/modules/trade/state/tradeStateFromUrlAtom.ts
@@ -1,0 +1,5 @@
+import { atom } from 'jotai'
+
+import { TradeRawState } from '../types/TradeRawState'
+
+export const tradeStateFromUrlAtom = atom<TradeRawState | null>(null)
